requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
braceexpand==0.1.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
import-expression==1.1.4
jishaku==2.5.1a461+g167b9ca.master
multidict==6.0.3
psutil==5.9.4
pycparser==2.21
PyNaCl==1.5.0
six==1.16.0
tabulate==0.9.0
yarl==1.8.2